ci: move lint code to a separate file
